requirements.txt
flask
requests
jsonpickle
grpcio
grpcio-tools
pillow